[REFACTOR] Extract control gateway interacting logic
In an attempt to bring more structure and control to the sprawling
`main.rs` file, create a `gateway_util` module to start consolidating
logic dealing with control gateway interactions.

`hab svc load`, `hab svc unload`, `hab svc start`, and `hab svc stop`
all follow the same basic pattern of "create a message, send it to the
gateway, and process the response". Rather than repeating the logic
four times, we extract it into the `send_message` function.

Ideally, we could channel all commands that interact with the control
gateway through the same logic, but the other commands are not
currently structured for that. Further refactoring may remedy this.
